Moved to double math for graphs (+speed), gain added, save/load done

+++ SpeakerEQDesigner/LogFunctionSeries.cs
@@ -9,45 +9,45 @@ using OxyPlot.Series;
 
 namespace SpeakerEQDesigner {
     class LogFunctionSeries : LineSeries {
-        private Func<decimal, decimal> func;
-        public Func<decimal, decimal> Function {
+        private Func<double, double> func;
+        public Func<double, double> Function {
             get { return func; }
             set { UpdateFunc(value, lowLim, upLim, step); }
         }
 
-        private decimal lowLim;
-        public decimal LowerLimit {
+        private double lowLim;
+        public double LowerLimit {
             get { return lowLim; }
             set { UpdateFunc(func, value, upLim, step); }
         }
 
-        private decimal upLim;
-        public decimal UpperLimit {
+        private double upLim;
+        public double UpperLimit {
             get { return upLim; }
             set { UpdateFunc(func, lowLim, value, step); }
         }
 
-        private decimal step;
-        public decimal Step {
+        private double step;
+        public double Step {
             get { return step; }
             set { UpdateFunc(func, lowLim, upLim, value); }
         }
 
-        public LogFunctionSeries(Func<decimal, decimal> f, decimal lowerLim, decimal upperLim, decimal logStep) {
+        public LogFunctionSeries(Func<double, double> f, double lowerLim, double upperLim, double logStep) {
             UpdateFunc(f, lowerLim, upperLim, logStep);
         }
 
-        private void UpdateFunc(Func<decimal, decimal> f, decimal lowerLim, decimal upperLim, decimal logStep) {
+        private void UpdateFunc(Func<double, double> f, double lowerLim, double upperLim, double logStep) {
             func = f;
             lowLim = lowerLim;
             upLim = upperLim;
             step = logStep;
 
             Points.Clear();
-            for (decimal x = lowerLim; x < upLim; x *= logStep) {
-                Points.Add(new DataPoint((double)x, (double)func(x)));
+            for (double x = lowerLim; x < upLim; x *= logStep) {
+                Points.Add(new DataPoint((double)x, func(x)));
             }
-            Points.Add(new DataPoint((double)upLim, (double)func(upLim)));
+            Points.Add(new DataPoint(upLim, func(upLim)));
         }
 
         protected override void UpdateData() {
